+++ sample/GatewayConnectedDeviceMgntAction/GatewayConnectedDeviceMgntAction.csproj
@@ -51,8 +51,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="ConectedDevice.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SampleGatewayConnectedDeviceManagementAction.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\IBMWIoTP\IBMWIoTP.csproj">
