+++ sample/GatewayConnectedDeviceMgntAction/SampleGatewayConnectedDeviceManagementAction.cs
@@ -16,126 +16,12 @@ using IBMWIoTP;
 
 namespace GatewayConnectedDeviceMgntAction
 {
-	public class ConectedDevice{
-		private IBMWIoTP.DeviceFirmware firmware = new DeviceFirmware();
-		public string deviceType {get;set;}
-		public string deviceId {get;set;}
-		public bool canReboot {get;set;}
-		public bool canReset {get;set;}
-		public bool isActionEnabled {get;set;}
-		public bool isFramewareEnabled {get;set;}
-		private GatewayManagement parentGateway;
-		
-		public ConectedDevice(string deviceType,string deviceId,GatewayManagement gateway){
-			this.deviceType = deviceType;
-			this.deviceId = deviceId;
-			this.parentGateway = gateway;
-		}
-		public void infoHandler(DeviceActionReq req){
-			var fields = req.d.fields;
-			for (int i = 0; i < fields.Length; i++) {
-				if(fields[i].field == "mgmt.firmware" ){
-					var update =  fields[i].value as IBMWIoTP.DeviceFirmware;
-					this.firmware.uri =update.uri;
-					this.firmware.verifier = update.verifier;
-					this.firmware.name = update.name;
-					this.firmware.state = update.state;
-					this.firmware.updatedDateTime = update.updatedDateTime;
-					this.firmware.updateStatus = update.updateStatus;
-					this.firmware.version = update.version;
-					break;
-				}
-			}
-			if(this.firmware != null)
-			{
-				parentGateway.sendResponse(req.reqId,204,"",this.deviceType,this.deviceId);
-			}
-		}
-		
-		public void downloadHandler (DeviceActionReq req){
-			int rc = DeviceManagement.RESPONSECODE_ACCEPTED;
-			string msg ="";
-			if(this.firmware.state != DeviceManagement.UPDATESTATE_IDLE){
-				rc = DeviceManagement.RESPONSECODE_BAD_REQUEST;
-				msg = "Cannot download as the device is not in idle state";
-				parentGateway.sendResponse(req.reqId,rc,msg,this.deviceType,this.deviceId);
-				return ;
-			}
-			parentGateway.sendResponse(req.reqId,rc,msg,this.deviceType,this.deviceId);
-			this.setState(DeviceManagement.UPDATESTATE_DOWNLOADING);
-			Console.WriteLine("Start downloading new Firmware form "+firmware.uri);
-			Thread.Sleep(2000);
-			//call your device to take all the action to download 
-			Console.WriteLine("completed Download");
-			this.setState(DeviceManagement.UPDATESTATE_DOWNLOADED);
-		
-		}
-		public void setState(int state){
-			if(this.firmware ==null)
-				return;
-			object[] field = new object[1];
-			field[0]= new {
-					        	field="mgmt.firmware",
-					        	value= new {
-					        		state=state
-					        	}
-			           };
-			var notify = new {
-				d = new {
-					fields=field
-					}
-				};
-			this.firmware.state = state;
-			string json = new System.Web.Script.Serialization.JavaScriptSerializer().Serialize(notify);
-			parentGateway.notify(this.deviceType,this.deviceId,json);
-		}
-		public void updatedHandler (DeviceActionReq req){
-			int rc = DeviceManagement.RESPONSECODE_ACCEPTED;
-			string msg ="";
-
-			if(this.firmware.state  != DeviceManagement.UPDATESTATE_DOWNLOADED){
-				rc = DeviceManagement.RESPONSECODE_BAD_REQUEST;
-				msg = "Firmware is still not successfully downloaded.";		
-				parentGateway.sendResponse(req.reqId,rc,msg,this.deviceType,this.deviceId);
-				return ;
-			}
-			parentGateway.sendResponse(req.reqId,rc,msg,this.deviceType,this.deviceId);
-			this.setUpdateState(DeviceManagement.UPDATESTATE_IN_PROGRESS);
-			Console.WriteLine("Start Updateting new Firmware ");
-			Thread.Sleep(2000);
-			Console.WriteLine("Updated new Firmware ");
-			this.setUpdateState(DeviceManagement.UPDATESTATE_SUCCESS);
-		
-		}
-		public void setUpdateState(int updateState){
-			if(this.firmware ==null)
-				return;
-			object[] field = new object[1];
-			field[0]= new {
-					        	field="mgmt.firmware",
-					        	value= new {
-					        		state=DeviceManagement.UPDATESTATE_IDLE,
-					        		updateStatus = updateState
-					        	}
-			           };
-			var notify = new {
-				d = new {
-					fields=field
-					}
-				};
-			
-			this.firmware.state = DeviceManagement.UPDATESTATE_IDLE;
-			this.firmware.updateStatus = updateState;
-			string json = new System.Web.Script.Serialization.JavaScriptSerializer().Serialize(notify);
-			parentGateway.notify(this.deviceType,this.deviceId,json);
-		}
-		
-	}
-	class Program
+	class SampleGatewayConnectedDeviceManagementAction
 	{
 		public static void Main(string[] args)
 		{
-
+        	Console.WriteLine("============================ IBM WatsonIoTP Sample ============================");
+	
 			bool isSync = false;
 			List<ConectedDevice> devices = new List<ConectedDevice>();
 			
@@ -145,9 +31,9 @@ namespace GatewayConnectedDeviceMgntAction
         	Console.WriteLine("2.Reset");
         	Console.WriteLine("3.Firmware download and update");
         	Console.WriteLine("4.Any key to exit");
-        	Console.Write("Please enter your choise :");
+        	Console.Write("Please enter your choice :");
         	int val = int.Parse(Console.ReadLine());
-        	
+	        	
 			DeviceInfo simpleDeviceInfo = new DeviceInfo();
 		    simpleDeviceInfo.description = "My device";
 		    simpleDeviceInfo.deviceClass = "My device class";
@@ -167,7 +53,7 @@ namespace GatewayConnectedDeviceMgntAction
 				Console.WriteLine("req Id:" + reqestId +"	Action:"+ action +" called");
 				if(action == "reboot"){
 					gwMgmtClient.sendResponse(reqestId,DeviceManagement.RESPONSECODE_ACCEPTED,"");
-
+	
 					Thread.Sleep(2000);
 					gwMgmtClient.disconnect();
 					
@@ -187,7 +73,7 @@ namespace GatewayConnectedDeviceMgntAction
 			gwMgmtClient.fwCallback += (string action , DeviceFirmware fw) => {
 				if(action == "download"){
 					gwMgmtClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADING);
-					Console.WriteLine("Start downloading new Firmware form "+fw.uri);
+					Console.WriteLine("Start downloading new Firmware from "+fw.uri);
 					Thread.Sleep(2000);
 					Console.WriteLine("completed Download");
 					gwMgmtClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADED);
@@ -195,7 +81,7 @@ namespace GatewayConnectedDeviceMgntAction
 				}
 				if(action == "update"){
 					gwMgmtClient.setUpdateState(DeviceManagement.UPDATESTATE_IN_PROGRESS);
-					Console.WriteLine("Start Updateting new Firmware ");
+					Console.WriteLine("Start Updating new Firmware ");
 					Thread.Sleep(2000);
 					Console.WriteLine("Updated new Firmware ");
 					gwMgmtClient.setUpdateState(DeviceManagement.UPDATESTATE_SUCCESS);
@@ -234,8 +120,8 @@ namespace GatewayConnectedDeviceMgntAction
 						}
 					}
 				}
-
-
+	
+	
 			};
 			
 			
@@ -271,21 +157,21 @@ namespace GatewayConnectedDeviceMgntAction
 			ele.typeId =  "testgw";
 			ele.deviceId= "1212";
 			deviceList[0] = ele;
-
-        	switch (val) {
-        		case 1 : client.InitiateDeviceManagementRequest("device/reboot",param,deviceList);
-        			break;
-        		case 2 : client.InitiateDeviceManagementRequest("device/factoryReset",param,deviceList);
-        			break;
-        		case 3 : client.InitiateDeviceManagementRequest("firmware/download",param,deviceList);
+	
+	        	switch (val) {
+	        		case 1 : client.InitiateDeviceManagementRequest("device/reboot",param,deviceList);
+	        			break;
+	        		case 2 : client.InitiateDeviceManagementRequest("device/factoryReset",param,deviceList);
+	        			break;
+	        		case 3 : client.InitiateDeviceManagementRequest("firmware/download",param,deviceList);
 						 Thread.Sleep(15000);
 						 client.InitiateDeviceManagementRequest("firmware/update",new IBMWIoTP.DeviceMgmtparameter[0],deviceList);
-        			break;
-        		default:
-        			gwMgmtClient.disconnect();
-        			break;
-        	}
-
+	        			break;
+	        		default:
+	        			gwMgmtClient.disconnect();
+	        			break;
+	        	}
+	
 		}
 	}
 }
