+++ sample/GatewayMgmtAction/SampleGatewayManagementAction.cs
@@ -13,12 +13,14 @@ using System;
 using IBMWIoTP;
 using System.Threading;
 
-namespace DeviceManagementAction
+namespace GatewayMgmtAction
 {
-	class Program
+	class SampleGatewayManagementAction
 	{
 		public static void Main(string[] args)
 		{
+        	Console.WriteLine("============================ IBM WatsonIoTP Sample ============================");
+			
 			string orgID = "";
 			string deviceType = "";
 			string deviceId = "";
@@ -26,20 +28,21 @@ namespace DeviceManagementAction
 			string authKey = "";
 			bool isSync = true;
 			
-			Console.WriteLine("Device Management Sample");
+			Console.WriteLine("Gateway Management Action Sample");
 		 	
 			Console.Write("Enter your org id :");
         	orgID = Console.ReadLine();
         	
-        	Console.Write("Enter your device type :");
+        	Console.Write("Enter your gateway type :");
         	deviceType = Console.ReadLine();
 
-        	Console.Write("Enter your device id :");
+        	Console.Write("Enter your gateway id :");
         	deviceId = Console.ReadLine();
 
         	Console.Write("Enter your auth key :");
         	authKey = Console.ReadLine();
 			
+
 			
 			DeviceInfo simpleDeviceInfo = new DeviceInfo();
 		    simpleDeviceInfo.description = "My device";
@@ -51,58 +54,54 @@ namespace DeviceManagementAction
 		    simpleDeviceInfo.serialNumber = "12345";
 		    simpleDeviceInfo.descriptiveLocation ="My device location";
 		    
-		    DeviceManagement	deviceClient = new DeviceManagement(orgID,deviceType,deviceId,authType,authKey,isSync);
-			deviceClient.deviceInfo = simpleDeviceInfo;
-			deviceClient.mgmtCallback += processMgmtResponse;
-			deviceClient.actionCallback += (string reqestId,string action)=>{
+		    GatewayManagement	gwMgmtClient = new GatewayManagement(orgID,deviceType,deviceId,authType,authKey,isSync);
+			gwMgmtClient.deviceInfo = simpleDeviceInfo;
+			gwMgmtClient.mgmtCallback += processMgmtResponse;
+			gwMgmtClient.actionCallback += (string reqestId,string action)=>{
 				Console.WriteLine("req Id:" + reqestId +"	Action:"+ action +" called");
 				if(action == "reboot"){
-					deviceClient.sendResponse(reqestId,DeviceManagement.RESPONSECODE_ACCEPTED,"");
+					gwMgmtClient.sendResponse(reqestId,DeviceManagement.RESPONSECODE_ACCEPTED,"");
 
 					Thread.Sleep(2000);
-					deviceClient.disconnect();
+					gwMgmtClient.disconnect();
 					
 					Console.WriteLine("disconnected");
 					Thread.Sleep(5000);
 					
 					Console.WriteLine("Re connected");	
-					deviceClient.connect();
+					gwMgmtClient.connect();
 					
-					deviceClient.manage(4000,true,true);
+					gwMgmtClient.managedGateway(4000,true,true);
 				}
 				if(action == "reset"){
-					deviceClient.sendResponse(reqestId,DeviceManagement.RESPONSECODE_FUNCTION_NOT_SUPPORTED,"");
+					gwMgmtClient.sendResponse(reqestId,DeviceManagement.RESPONSECODE_FUNCTION_NOT_SUPPORTED,"");
 				}
 		
 			};
-			deviceClient.fwCallback += (string action , DeviceFirmware fw)=>{
+			gwMgmtClient.fwCallback += (string action , DeviceFirmware fw)=>{
 				if(action == "download"){
-					deviceClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADING);
-					Console.WriteLine("Start downloading new Firmware form "+fw.uri);
+					gwMgmtClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADING);
+					Console.WriteLine("Start downloading new Firmware from "+fw.uri);
 					Thread.Sleep(2000);
 					Console.WriteLine("completed Download");
-					deviceClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADED);
+					gwMgmtClient.setState(DeviceManagement.UPDATESTATE_DOWNLOADED);
 				
 				}
 				if(action == "update"){
-					deviceClient.setUpdateState(DeviceManagement.UPDATESTATE_IN_PROGRESS);
-					Console.WriteLine("Start Updateting new Firmware ");
+					gwMgmtClient.setUpdateState(DeviceManagement.UPDATESTATE_IN_PROGRESS);
+					Console.WriteLine("Start Updating new Firmware ");
 					Thread.Sleep(2000);
 					Console.WriteLine("Updated new Firmware ");
-					deviceClient.setUpdateState(DeviceManagement.UPDATESTATE_SUCCESS);
+					gwMgmtClient.setUpdateState(DeviceManagement.UPDATESTATE_SUCCESS);
 				}
 			};
-			deviceClient.connect();
-			deviceClient.subscribeCommand("testcmd", "json", 2);
-           	deviceClient.commandCallback += processCommand;
+			gwMgmtClient.connect();
 			Console.WriteLine("Manage");
-			deviceClient.manage(4000,true,true);
-			Console.WriteLine("Set Location");
-			deviceClient.setLocation(77.5667,12.9667, 0,10);
-
-			//Console.Write("Press any key to exit . . . ");
+			gwMgmtClient.managedGateway(4000,true,true);
+		
+//			Console.Write("Press any key to exit . . . ");
 			Console.ReadKey();
-//			deviceClient.disconnect();
+//			gwMgmtClient.disconnect();
 			
 			
 		}
@@ -110,8 +109,7 @@ namespace DeviceManagementAction
 			Console.WriteLine("req Id:" + reqestId +"	responseCode:"+ responseCode);
 		}
 		
-        public static void processCommand(string cmdName, string format, string data) {
-             Console.WriteLine("Sample Device Client : Sample Command " + cmdName + " " + "format: " + format + "data: " + data);
-        }
+        
 	}
+	
 }
